Refactored store related things into separate package
