+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/util/JsonObjectWrapper.java
@@ -1,0 +1,31 @@
+/*
+ *
+ *  * Copyright 2021 Crown Copyright
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *     http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package uk.gov.gchq.gaffer.gaas.util;
+
+import java.util.Collections;
+import java.util.Map;
+
+public final class JsonObjectWrapper {
+    private JsonObjectWrapper() {
+    }
+
+    public static <Gaffer> Map<String, Gaffer> withLabel(final String label, final Gaffer wrappedObject) {
+        return Collections.singletonMap(label, wrappedObject);
+    }
+}
