+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/factories/StoreTypeFactoryTest.java
@@ -19,6 +19,12 @@ package uk.gov.gchq.gaffer.gaas.factories;
 import org.junit.jupiter.api.Test;
 import uk.gov.gchq.gaffer.common.model.v1.GafferSpec;
 import uk.gov.gchq.gaffer.gaas.model.GaaSCreateRequestBody;
+import uk.gov.gchq.gaffer.gaas.stores.AbstractStoreTypeBuilder;
+import uk.gov.gchq.gaffer.gaas.stores.AccumuloStoreType;
+import uk.gov.gchq.gaffer.gaas.stores.FederatedStoreType;
+import uk.gov.gchq.gaffer.gaas.stores.MapStoreType;
+import uk.gov.gchq.gaffer.gaas.stores.ProxyStoreType;
+import uk.gov.gchq.gaffer.gaas.stores.StoreType;
 import uk.gov.gchq.gaffer.gaas.utilities.UnitTest;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
